+++ src/services/HomotopyTreeUtils.jl
@@ -44,10 +44,10 @@ function getKernelLeaf(
 
     # FIXME refactor in transit, use HR{,,,A} to extract leaf and minor relation
     # TODO consolidate with uniBW()
-    cv_ = if 1 == SparseArrays.nnz(hode.minors_detail)
-        hode.minors_detail[1]
+    cv_ = if 1 == SparseArrays.nnz(hode.trailing_forms)
+        hode.trailing_forms[1]
     else
-        hode.minors_detail[i]
+        hode.trailing_forms[i]
     end
 
     hr = hode.reprkind
@@ -57,7 +57,7 @@ function getKernelLeaf(
     # FIXME, hack before reworking partials to common trait -- 
     #  partial and partl_cb elsewhere assumed to travel together, not recreated post-hoc
     _, _, partl_cb = getManifoldPartial(getManifold(hode), _tuple(partial), mn)
-    return getReprType(hr)(mn, cv_; partial, partl_cb)
+    return _vanillareprT(getFormKind(hr))(mn, cv_; partial, partl_cb)
 end
 
 """
@@ -73,7 +73,7 @@ function getKernelLeafAsTreeKer(
     idx::Int,
     permuted::Bool = false,
 ) where {H, P}
-    reprT = getReprType(mtr.reprkind)
+    reprT = getFormKind(mtr.reprkind) |> _vanillareprT
     partial = getPartial(mtr.reprkind)
     mani = getManifold(mtr)
     lidx = (idx - 1) % Npts(mtr) + 1
@@ -107,20 +107,21 @@ function getKernelTree(
     cov_continuation::Bool = false,
 ) where {H, P}
     #
+
     N = Npts(hode)
     partial = getPartial(hode)
-    reprT = getReprType(hode.reprkind)
+    reprT = getFormKind(hode) |> _vanillareprT
     # BinaryTree (BT) index goes from root=1 to largest leaf 2*N
     return if isassigned(hode, currIdx) && !isLeaf_BTLabel(hode, currIdx)
         # cov_continuation correction so that we may build trees with sensible convariance to bandwidth transition from root to leaf
-        μ = hode.majors_element[currIdx]
+        μ = hode.principal_elements[currIdx]
             # FIXME, hack before reworking partials to common trait -- 
             #  partial and partl_cb elsewhere assumed to travel together, not recreated post-hoc
             _, _, partl_cb = getManifoldPartial(getManifold(hode), _tuple(partial), μ)
         raw_ker = reprT(
             μ, 
-            hode.majors_detail[currIdx], 
-            hode.majors_coeff[currIdx];
+            hode.principal_forms[currIdx],
+            hode.principal_coeffs[currIdx];
             partial, partl_cb
         )
         if cov_continuation